[refactor] Finished moving commands from dmm.ts
