+++ dmm.ts
@@ -118,113 +118,3 @@ async function constructModulesDataFromDeps(
   }
   return modules;
 }
-
-/**
- * Main logic for purposes of this module.
- */
-export const purposes: { [key: string]: Function } = {
-  "update": async function (modulesForPurpose: string[]) {
-    // Create objects for each dep, with its name and version
-    const purpose = "update";
-    const modules = await constructModulesDataFromDeps(
-      modulesForPurpose,
-      purpose,
-    );
-    if (modules === false || typeof modules === "boolean") {
-      console.error(
-        colours.red("Modules specified do not exist in your dependencies."),
-      );
-      Deno.exit(1);
-      return;
-    }
-
-    // Check for updates and rewrite `deps.ts` if needed
-    console.info("Checking if your modules can be updated...");
-    const usersWorkingDir: string = Deno.realPathSync(".");
-    let depsWereUpdated = false;
-    let depsContent: string = decoder.decode(
-      Deno.readFileSync(usersWorkingDir + "/deps.ts"),
-    ); // no need for a try/catch. The user needs a deps.ts file
-    modules.forEach((module) => {
-      // only re-write modules that need to be updated
-      if (module.importedVersion === module.latestRelease) {
-        return;
-      }
-      depsContent = depsContent.replace(
-        "std@" + module.importedVersion + "/" + module.name,
-        "std@" + module.latestRelease + "/" + module.name,
-      );
-      console.info(
-        colours.green(
-          module.name + " was updated from " + module.importedVersion + " to " +
-            module.latestRelease,
-        ),
-      );
-      depsWereUpdated = true;
-    });
-
-    // Re-write the file
-    Deno.writeFileSync(
-      usersWorkingDir + "/deps.ts",
-      new TextEncoder().encode(depsContent),
-    );
-
-    // And if none were updated, add some more logging
-    if (!depsWereUpdated) {
-      console.info(colours.green("Everything is already up to date"));
-    }
-  },
-  "info": async function (modulesForPurpose: string[]) {
-    if (modulesForPurpose.length === 0 || modulesForPurpose.length > 1) {
-      console.error(
-        colours.red(
-          "Specify a single module to get information on. See --help",
-        ),
-      );
-      Deno.exit(1);
-    }
-    const moduleToGetInfoOn = modulesForPurpose[0];
-    const purpose = "info";
-    const database = await getDenoLandDatabase();
-    const stdResponse = await fetch(
-      "https://github.com/denoland/deno/tree/master/std/" + moduleToGetInfoOn,
-    );
-    const isStd = stdResponse.status === 200;
-    const isThirdParty = typeof database[moduleToGetInfoOn] === "object";
-    if (!isStd && !isThirdParty) {
-      console.error(
-        colours.red("No module was found with " + moduleToGetInfoOn),
-      );
-      Deno.exit(1);
-    }
-    const name = moduleToGetInfoOn;
-    let description;
-    let denoLandUrl;
-    let gitHubUrl;
-    let latestVersion;
-    if (isStd) {
-      latestVersion = await getStdLatestVersion();
-      description = "Cannot retrieve descriptions for std modules";
-      denoLandUrl = "https://deno.land/std@" + latestVersion + "/" +
-        name;
-      gitHubUrl = "https://github.com/denoland/deno/tree/master/std/" + name;
-    }
-    if (isThirdParty) {
-      const databaseModule = database[moduleToGetInfoOn];
-      description = databaseModule.desc;
-      gitHubUrl = "https://github.com/" + databaseModule.owner + "/" +
-        databaseModule.repo;
-      const res = await fetch(gitHubUrl + "/releases/latest");
-      const splitUrl: string[] = res.url.split("/");
-      latestVersion = splitUrl[splitUrl.length - 1];
-      denoLandUrl = "https://deno.land/x/" + name + "@" + latestVersion;
-    }
-    const importLine = "import * as " + name + ' from "' + denoLandUrl + '";';
-    console.info(
-      "\n" +
-        `Information on ${name}\n\n  - Name: ${name}\n  - Description: ${description}\n  - deno.land Link: ${denoLandUrl}\n  - GitHub Repository: ${gitHubUrl}\n  - Import Statement: ${importLine}\n  - Latest Version: ${latestVersion}` +
-        "\n",
-    );
-    Deno.exit();
-  },
-};
