+++ src/utils.ts
@@ -1,4 +1,4 @@
-function standardiseVersion (importedVersion: string, latestVersion: string): string {
+export function standardiseVersion (importedVersion: string, latestVersion: string): string {
   const importedVersionHasV = importedVersion.indexOf("v") === 0;
   const latestVersionHasV = latestVersion.indexOf("v") === 0;
 
@@ -34,23 +34,20 @@ async function getDenoLandDatabase(): Promise<DenoLandDatabase> {
   const denoDatabase: DenoLandDatabase = await res.json();
   return denoDatabase;
 }
-const denoLandDatabase: DenoLandDatabase = await getDenoLandDatabase();
+export const denoLandDatabase: DenoLandDatabase = await getDenoLandDatabase();
 
 /**
  * @description
  * Fetches the latest std version
  *
- * @param {string} importedVersion To check if the returned val needs a "v" or not
- *
  * @returns {Promise<string>} eg "0.57.0"
  */
-export async function getLatestStdRelease(importedVersion: string): Promise<string> {
+export async function getLatestStdRelease(): Promise<string> {
   const res = await fetch(
       "https://raw.githubusercontent.com/denoland/deno_website2/master/deno_std_versions.json",
   );
   const versions: string[] = await res.json();
   let latestVersion = versions[0];
-  latestVersion = standardiseVersion(importedVersion, latestVersion)
   return latestVersion;
 }
 
@@ -60,12 +57,11 @@ export async function getLatestStdRelease(importedVersion: string): Promise<stri
  * Achieves this by reading Deno's `database.json` to get the repository name
  * and owner then sending a fetch request.
  *
- * @param {string} importedVersion So returned value can standardise in having "v" or not
  * @param {string} name Module name
  *
  * @returns {Promise<string>} The latest version.
  */
-export async function getLatestThirdPartyRelease(importedVersion: string, name: string): Promise<string> {
+export async function getLatestThirdPartyRelease(name: string): Promise<string> {
   const owner = denoLandDatabase[name].owner;
   const repo = denoLandDatabase[name].repo;
   const res = await fetch(
@@ -74,6 +70,5 @@ export async function getLatestThirdPartyRelease(importedVersion: string, name: 
   const url = res.url;
   const urlSplit = url.split("/");
   let latestRelease = urlSplit[urlSplit.length - 1];
-  latestRelease = standardiseVersion(importedVersion, latestRelease)
   return latestRelease;
 }
